data-loader-service/requirements.txt
psycopg2-binary==2.9.9
minio==7.2.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
requests==2.32.3
